requirements.txt
Flask>=3.0.2,<3.1
Flask-Cors>=4.0.0,<4.1
Flask-SQLAlchemy>=3.1.1,<3.2
requests>=2.31.0,<2.32
python-whois>=0.9.3,<1.0
pandas>=2.2.1,<2.3
aiohttp>=3.9.3,<4.0
keras==3.1.1
tensorflow==2.16.1